processor/src/test/kotlin/com/octo/workerdecorator/processor/generator/java/JavaMutableWeakExecutorGeneratorTest.kt
package com.octo.workerdecorator.processor.generator.java

import com.octo.workerdecorator.processor.test.CompilationAwareTest
import org.assertj.core.api.Assertions.assertThat
import org.junit.Test

class JavaMutableWeakExecutorGeneratorTest : CompilationAwareTest() {

    @Test
    fun `it generates the wanted java file`() {
        // Given
        val generator = JavaMutableWeakExecutorGenerator()
        val document = simpleInterfaceFixture()

        // When
        val result = generator.generate(document)

        // Then
        val targetContent = readResource("ExpectedSimpleInterfaceWeakMutableWorkerDecoration.java")
        assertThat(result).isEqualTo(targetContent)
    }
}